Move: getUTXOsByAddress to externalApis
